Refactor everything to use uuids
We were using triple-identifiers for settings (extension-name,
bundle-key, setting-key) and tuple-identifiers for
bundles (extension-name, bundle-key). As it turns out this is not
working out well for referencing e.g. settings from within other
settings, which is required for adding permissions as settings. Thus we
changed our protobuf definition to have uuids insteads of complex
identifiers.
Along with changing to identifiers we needed to rewrite our persistence
implementation. GRPC tests are broken and will be fixed in one of the
next commits.